feat: log anthropic calls
Prior to this commit, there were times where I wanted to review the
output of the model after revising and finalising the commit, but it was
not possible to do so as an out-of-band operation.

This change adds a feature to log the output of the model to a file
when ANTHROPIC_LOG_DIR is set. The log file is named after the tree
hash of the commit and contains the exact contents of the file as it
was presented to the user. Any errors that occur during the logging
process are printed to stderr and the process continues as normal.